requirements.txt
flask==3.0.2
humanize==4.10.0
flask-caching==2.3.0
websocket-client==1.8.0
gunicorn==22.0.0
bcrypt==4.1.2
requests==2.31.0